// src/services/openapi/models/apigen.types.ts
export interface JavaProperties {
  'group-id': string;
  'artifact-id': string;
}

export interface ApigenProject {
  name: string;
  description?: string;
  version: string;
  partial?: boolean;
  'data-driver'?: string;
  'java-properties'?: JavaProperties;
}

export interface ApigenModelAttributeRelationalPersistence {
  column?: string;
  'foreign-column'?: string;
  'primary-key'?: boolean;
  unique?: boolean;
  autogenerated?: boolean;
  'join-table'?: string;
  'join-column'?: string;
  'inverse-join-column'?: string;
  sequence?: string;
}

export interface ApigenModelAttributeValidation {
  type: string;
  min?: number;
  max?: number;
  regex?: string;
  value?: string;
  integer?: number;
  fraction?: number;
  inclusive?: boolean;
}

export interface ApigenModelAttribute {
  name: string;
  type: string;
  'items-type'?: string;
  'relational-persistence'?: ApigenModelAttributeRelationalPersistence;
  validations?: ApigenModelAttributeValidation[];
}

export interface ApigenModelRelationalPersistence {
  table: string;
}

export interface ApigenModel {
  'relational-persistence'?: ApigenModelRelationalPersistence;
  attributes: ApigenModelAttribute[];
}

export interface ApigenBinding {
  model: string;
  [paramName: string]: string;
}

export interface ApigenMapping {
  model?: string;
  field?: string;
  method?: string;
}

export interface ApigenExtensions {
  'x-apigen-project'?: ApigenProject;
  'x-apigen-models'?: Record<string, ApigenModel>;
  'x-apigen-binding'?: ApigenBinding;
  'x-apigen-mapping'?: ApigenMapping;
}